Move KVM virtio-blk code into blk.go file.
